Decompile RIC StopwatchCrashDoneSparkle
