+++ config/splat.us.ric.yaml
@@ -53,8 +53,6 @@ segments:
       - [0x1A9E8, rodata]
       - [0x1AB04, rodata]
       - [0x1AB68, .rodata, 32324]
-      - [0x1ABB0, rodata]
-      - [0x1ABC8, .rodata, 345EC]
       - [0x1AC50, rodata]
       - [0x1AC60, c, 1AC60]
       - [0x1F348, c, 1F348]
@@ -65,6 +63,5 @@ segments:
       - [0x24788, c, 24788]
       - [0x26C84, c, 26C84]
       - [0x32324, c, 32324]
-      - [0x345EC, c, 345EC]
       - [0x37B64, sbss]
   - [0x39A58]
