+++ config/symbols.us.dra.txt
@@ -19,7 +19,6 @@ g_Vram_D_800ACDD8 = 0x800ACDD8;
 g_Vram_D_800ACDE0 = 0x800ACDE0;
 g_Vram_D_800ACDE8 = 0x800ACDE8;
 g_Vram_D_800ACDF0 = 0x800ACDF0;
-g_DraEntityTbl = 0x800AD0C4;
 D_800B0608 = 0x800B0608;
 aPbav = 0x800B107C;
 aPbav_0 = 0x800B407C;
